Begin making most of the x86 projects to track at Android-Generic forks
After spending some time making forks at AG github and gitlab, we are
ready to track there. Almost all of the projects at 02-android-x86.xml
will now move to AG forks with some exceptions. This is to avoid confusion
when looking at the list in the manifest, also it will help Android-Generic
itself as nearly all the required projects are in AG Github & Gitlab already.

Signed-off-by: hmtheboy154 <buingoc67@gmail.com>